Refactoring: generic abstract types

+++ src/test/java/com/gmail/uprial/customcreatures/schema/HValueTest.java
@@ -40,7 +40,7 @@ public class HValueTest extends TestConfigBase {
     @Test
     public void testSimpleIntValue() throws Exception {
         //noinspection ConstantConditions
-        assertEquals(42, getIntFromConfig(getPreparedConfig("i: 42"),
+        assertEquals(42, (int)getIntFromConfig(getPreparedConfig("i: 42"),
                 getParanoiacCustomLogger(), "i", "i of handler", "x").getValue());
     }
 
@@ -59,7 +59,7 @@ public class HValueTest extends TestConfigBase {
     @Test
     public void testRandomIntValue() throws Exception {
         //noinspection ConstantConditions
-        assertEquals(2, getIntFromConfig(getPreparedConfig(
+        assertEquals(2, (int)getIntFromConfig(getPreparedConfig(
                 "i:",
                 " type: random",
                 " min: 2",
@@ -97,5 +97,4 @@ public class HValueTest extends TestConfigBase {
         assertEquals(null, getIntFromConfig(getPreparedConfig(""),
                 getDebugFearingCustomLogger(), "i", "i of handler", "x"));
     }
-
 }
